+++ gem/config/pagy.rb
@@ -9,7 +9,7 @@
 # Examples:
 #
 # Pagy::OPTIONS[:limit]     = 10     # Limit the items per page
-# Pagy::OPTIONS[:max_limit] = 100    # The client is allowed to request a limit up to 100
+# Pagy::OPTIONS[:client_limit] = 100    # The client is allowed to request a limit up to 100
 # Pagy::OPTIONS[:jsonapi]   = true   # Use JSON:API compliant URLs
 
 Pagy::OPTIONS.freeze